+++ packages/cli/src/render-ts/types.ts
@@ -1,6 +1,10 @@
+import { SchemaAbiType, StaticAbiType } from "@latticexyz/schema-type";
+
 export interface RecsV1TableOptions {
   tables: {
     tableName: string;
+    keySchema: Record<string, StaticAbiType>;
+    valueSchema: Record<string, SchemaAbiType>;
     fields: {
       recsTypeString: string;
       name: string;