+++ packages/store/src/StoreRead.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 
-import { IStoreRead, IStoreHook } from "./IStore.sol";
+import { IStoreRead } from "./IStore.sol";
 import { StoreCore } from "./StoreCore.sol";
 import { Schema } from "./Schema.sol";
 
@@ -60,6 +60,4 @@ contract StoreRead is IStoreRead {
   ) public view virtual returns (bytes memory) {
     return StoreCore.getFieldSlice(tableId, key, schemaIndex, schema, start, end);
   }
-
-  function isStore() public view {}
 }