+++ packages/store/ts/index.ts
@@ -2,5 +2,6 @@
 // (library neither creates nor extends MUDCoreContext when imported)
 export * from "./config";
 
+export * from "./common";
 export * from "./storeEvents";
 export * from "./storeEventsAbi";