+++ templates/phaser/packages/contracts/test/CounterTest.t.sol
@@ -2,13 +2,13 @@
 pragma solidity >=0.8.0;
 
 import "forge-std/Test.sol";
-import { MudV2Test } from "@latticexyz/std-contracts/src/test/MudV2Test.t.sol";
+import { MudTest } from "@latticexyz/store/src/MudTest.sol";
 import { getKeysWithValue } from "@latticexyz/world/src/modules/keyswithvalue/getKeysWithValue.sol";
 
 import { IWorld } from "../src/codegen/world/IWorld.sol";
 import { Counter, CounterTableId } from "../src/codegen/Tables.sol";
 
-contract CounterTest is MudV2Test {
+contract CounterTest is MudTest {
   IWorld public world;
 
   function setUp() public override {